% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_projection.R
\name{adjust_projection_catcheco_jabba}
\alias{adjust_projection_catcheco_jabba}
\title{Calculate catch for projections based on JABBA outputs}
\usage{
adjust_projection_catcheco_jabba(fmsy_melted, Bt)
}
\arguments{
\item{fmsy_melted}{Melted data frame of Fmsy objects.}

\item{Bt}{Biomass in terminal model year from JABBA outputs.}
}
\value{
Adjusted projection values (i.e., FMSY).
}
\description{
Calculate catch for projections based on JABBA outputs
}
