% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_projection.R
\name{adjust_projection_dbsra}
\alias{adjust_projection_dbsra}
\title{Adjust TAC based on DB-SRA outputs}
\usage{
adjust_projection_dbsra(tac, soi, Bt_BMSY)
}
\arguments{
\item{soi}{Status of indicator over time.}

\item{Bt_BMSY}{Biomass over BMSY estimates from DB-SRA outputs.}

\item{TAC}{A vector of TAC estimates from DB-SRA outputs.}
}
\value{
Adjusted projection values (i.e., TAC).
}
\description{
Adjust TAC based on DB-SRA outputs
}
