% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_projection.R
\name{adjust_projection_jabba}
\alias{adjust_projection_jabba}
\title{Adjust FMSY based on JABBA outputs}
\usage{
adjust_projection_jabba(FMSY, soi, Bt_BMSY)
}
\arguments{
\item{FMSY}{A vector of FMSY estimates from JABBA outputs.}

\item{soi}{Status of indicator over time.}

\item{Bt_BMSY}{Biomass over BMSY estimates from JABBA outputs.}
}
\value{
Adjusted projection values (i.e., FMSY).
}
\description{
Adjust FMSY based on JABBA outputs
}
