% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_status_of_indicator.R
\name{calc_soi}
\alias{calc_soi}
\title{Calculate status of an indicator over time}
\usage{
calc_soi(indicator_data, slope)
}
\arguments{
\item{indicator_data}{A time series data of an indicator (e.g., environmental, ecological, and economical indicators).}

\item{slope}{A single slope value from first-order linear regression analysis.}
}
\value{
Status of an indicator over time.
}
\description{
Calculate status of an indicator over time
}
