% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_projections.R
\name{compare_projections_dbsra}
\alias{compare_projections_dbsra}
\title{Compare projections from different DB-SRA cases}
\usage{
compare_projections_dbsra(
  case_name,
  projection_output_data,
  projection_year,
  model_year,
  lag = 0,
  amo,
  pdsi,
  bassB,
  menhadenC,
  menhadenE,
  menhadenCPUE,
  bassCPUE,
  herringCPUE,
  menhadenV,
  indices
)
}
\arguments{
\item{case_name}{name of a case.}

\item{projection_output_data}{Output data frame from different DB-SRA cases.}

\item{projection_year}{A vector of projection years.}

\item{model_year}{A vector of model years.}

\item{lag}{Lag between biomass of menhaden-like species and indices.}

\item{amo}{AMO data.}

\item{pdsi}{Palmer drought severity index data.}

\item{bassB}{Biomass of Striped bass data.}

\item{menhadenC}{Catch of menhaden-like species from the EwE operating model.}

\item{menhadenE}{Fishing effort of menhaden-like species from the EwE operating model.}

\item{menhadenCPUE}{Catch per unit effort of menhaden-like species from the EwE operating model.}

\item{bassCPUE}{Catch per unit effort of bass-like species from the EwE operating model.}

\item{herringCPUE}{Catch per unit effort of herring-like species from the EwE operating model.}

\item{menhadenV}{Ex-vessel value of menhaden-like species.}

\item{indices}{Indices will be used in the analysis.}
}
\value{
A data list that includes slope values from linear regression models (lm_slope), status of indicator output (soi_data), adjusted TAC output (tac_data), reshaped status of indicator output (soi_data_melt), reshaped adjusted TAC output (tac_data_melt), and median adjusted tac over years (tac_data_melt_median).
}
\description{
Compare projections from different DB-SRA cases
}
