% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biodata.R
\name{create_biodata}
\alias{create_biodata}
\title{Create biological data}
\usage{
create_biodata(
  nsex = 1,
  narea = 1,
  ages,
  years,
  length_bin,
  mid_length_bin,
  nbin,
  bin_width,
  length_CV,
  annual_weight_path,
  monthly_weight_path,
  species,
  species_labels,
  skip_nrows,
  lw_a = 0.01,
  lw_b = 3,
  k,
  maturity_at_age,
  natural_mortality_at_age
)
}
\arguments{
\item{nsex}{Number of sexes of the species.}

\item{narea}{Number of areas.}

\item{ages}{A vector of ages.}

\item{years}{A vector of years for extracting EwE data.}

\item{length_bin}{A vector of length bins.}

\item{mid_length_bin}{A vector of middle point of the length bins.}

\item{nbin}{Total number of bins.}

\item{bin_width}{Width of bins.}

\item{length_CV}{CV about length.}

\item{annual_weight_path}{A character string shows path to Ecosim annual weight-at-age output file is located.}

\item{monthly_weight_path}{A character string shows path to Ecosim monthly weight-at-age output file is located.}

\item{species}{The species to create fishery for. A vector of column names from the catch_annual.csv.}

\item{species_labels}{Customized names for species columns.}

\item{skip_nrows}{integer; the number of lines of the data file to skip before reading data.}

\item{lw_a}{Length-weight coefficients a. W=a*L^b. Default EwE value is 0.01.}

\item{lw_b}{Length-weight coefficients b. W=a*L^b. Default EwE value is 3.}

\item{k}{Carrying capacity parameter of the von Bertalanffy model.}

\item{maturity_at_age}{A vector of maturity-at-age value or time-varying maturity-at-age matrix with rows present years and columns represent ages.}

\item{natural_mortality_at_age}{A vector of natural mortality-at-age value or time-varying natural mortality-at-age matrix with rows present years and columns represent ages.}
}
\value{
A data list that includes biological data for set up a stock assessment.
}
\description{
This function creates a data list that contains remaining biological data for a stock assessment
}
\details{
The function takes EwE input or output data to create weight-at-age matrix, maturity-at-age matrix,
}
\seealso{
Other create_data functions: 
\code{\link{create_fishery}()}
}
\concept{create_data functions}
