% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishery.R, R/create_survey.R
\name{create_fishery}
\alias{create_fishery}
\title{Create fishery from Ecopath with Ecosim (EwE) output data}
\usage{
create_fishery(
  file_path,
  skip_nrows,
  species,
  species_labels,
  years,
  fleet_num,
  selectivity = NULL,
  CV,
  sample_num,
  waa_path
)

create_fishery(
  file_path,
  skip_nrows,
  species,
  species_labels,
  years,
  fleet_num,
  selectivity = NULL,
  CV,
  sample_num,
  waa_path
)
}
\arguments{
\item{file_path}{A character string shows path to Ecosim biomass output file (e.g., biomass_monthly.csv) is located.}

\item{skip_nrows}{interger; the number of lines of the data file to skip before reading data.}

\item{species}{The species to create fishery for. A vector of column names from the catch_annual.csv.}

\item{species_labels}{Customized names for speicies columns.}

\item{years}{A vector of years of the survey.}

\item{fleet_num}{Fleet number.}

\item{selectivity}{Selectivity-at-age. A list of dataframe defining selectivity-at-age for a species over time. Each list represent a survey. For each dataframe, columns represent age classes and rows represent years.}

\item{CV}{CV of survey indices. A vector of CVs for individual surveys.}

\item{sample_num}{Annual sample size. A vector of sample sizes for indiviudal surveys.}

\item{waa_path}{A character string shows path to Ecosim weight output file (e.g., weight_monthly.csv) is located.}

\item{survey_num}{Number of surveys.}
}
\value{
A data list that includes catch and composition data from the EwE operating model, and observed catch and composition data for each fleet. Each fleet has time series catch with lognormal error and composition data with sampling error.

A data list that includes survey indices and composition data from the EwE operating model, and observed survey indices and composition data for each survey.
Each survey has time series catch with lognormal error and composition data with sampling error.
}
\description{
This function creates fishery observations for an Ecosim scenario.

This function creates survey observations for an Ecosim scenario.
}
\details{
The function takes fishery catch data from an Ecosim scenario.
The catch data are read in from Ecosim catch_annual.csv and the catch data are in biomass.
The function reads in total catch-at-age data of a specific species by age classes and time.
EwE already applies a fishery selectivity internally, so no need to reallocate catch if there is only one fleet for the fishery.

The function takes biomass data from an Ecosim scenario.
The biomass data are read in from Ecosim biomass_annual.csv or biomass_monthly.csv.
The function reads in total biomass-at-age data of a specific species by age classes and time.
}
\concept{create_data functions}
