% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey.R
\name{create_survey}
\alias{create_survey}
\title{Create survey from Ecopath with Ecosim (EwE) output data}
\usage{
create_survey(
  file_path,
  skip_nrows,
  species,
  species_labels,
  years,
  survey_num,
  survey_time,
  selectivity = NULL,
  catchability,
  CV,
  sample_num,
  waa_path
)
}
\arguments{
\item{file_path}{A character string shows path to Ecosim biomass output file (e.g., biomass_monthly.csv) is located.}

\item{skip_nrows}{integer; the number of lines of the data file to skip before reading data.}

\item{species}{The species to create fishery for. A vector of column names from the catch_annual.csv.}

\item{species_labels}{Customized names for species columns.}

\item{years}{A vector of years for extracting EwE data.}

\item{survey_num}{Number of surveys.}

\item{survey_time}{Survey years and months. A list of data frame for defining survey time. Each list represent a survey. For each data frame, columns represent year and month.}

\item{selectivity}{Selectivity-at-age. A list of data frame defining selectivity-at-age for a species over time. Each list represent a survey. For each data frame, columns represent age classes and rows represent years.}

\item{CV}{Coefficient of variation of survey indices. A vector of CVs for individual surveys.}

\item{sample_num}{Annual sample size. A vector of sample sizes for individual surveys.}

\item{waa_path}{A character string shows path to Ecosim weight output file (e.g., weight_monthly.csv) is located.}

\item{cathability}{Catchability of survey indices. A vector of catchability values for individual surveys.}
}
\value{
A data list that includes survey indices and composition data from the EwE operating model, and observed survey indices and composition data for each survey. Each survey has time series catch with lognormal error and composition data with sampling error.
}
\description{
This function creates survey observations for an Ecosim scenario.
}
\details{
The function takes biomass data from an Ecosim scenario.
The biomass data are read in from Ecosim biomass_annual.csv or biomass_monthly.csv.
The function reads in total biomass-at-age data of a specific species by age classes and time.
}
