% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cumd_norm}
\alias{cumd_norm}
\title{Cumulative normal distribution}
\usage{
cumd_norm(x)
}
\arguments{
\item{x}{A normalized "Z" score value (subtract the mean and divide by the standard deviation).}
}
\value{
Probability that of an observation will exceed the argument.
}
\description{
Cumulative normal distribution
}
