% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downlaod_envrionment_data.R
\name{download_amo}
\alias{download_amo}
\title{Download AMO data}
\usage{
download_amo(url, years, output_dir, scale_range = c(0, 2))
}
\arguments{
\item{url}{url for downloading data.}

\item{years}{a vector of years.}

\item{output_dir}{path to output csv file.}

\item{scale_range}{output range (numeric vector of length two).}
}
\description{
It downloads data from https://psl.noaa.gov/data/timeseries/AMO/, reshape the data to one column, and save the data to a specified location. It returns a dataframe with 2 columns. One column contains raw values and the other column contains scaled values with specified output range.
}
