% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_jabba.R
\name{generate_jabba}
\alias{generate_jabba}
\title{Generate frequently used input data for JABBA}
\usage{
generate_jabba(
  assessment_name,
  output_dir,
  sa_data,
  BmsyK = 0.4,
  model_year,
  projection_year,
  effort_data = NULL,
  tacs = NULL
)
}
\arguments{
\item{assessment_name}{name of the assessment}

\item{output_dir}{A file path to a directory where the assessment output will be saved.}

\item{sa_data}{Simulated stock assessment input data from Rscript/simulation.R.}

\item{BmsyK}{Inflection point of the surplus production curve. Default value is 0.4.}

\item{model_year}{A vector of years for model fitting.}

\item{projection_year}{A vector of years for projection.}

\item{effort_data}{A vector of effort data.}

\item{tacs}{A vector of total allowable catch values for projection.}
}
\value{
A list of initial input values
}
\description{
Generate frequently used input data for JABBA
}
\examples{
\dontrun{
generate_jabba(output_dir = here::here("data", "data_moderate"))
}
}
