% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ss3.R
\name{generate_ss3}
\alias{generate_ss3}
\title{Generate frequently used input data for Stock Synthesis 3}
\usage{
generate_ss3(
  file_path,
  r0,
  r0_phase = 1,
  steepness,
  sigmar,
  projection_f,
  projection_catch = NULL,
  sa_data,
  model_year,
  projection_year,
  use_depletion = FALSE,
  depletion_ratio = NULL,
  initial_equilibrium_catch = TRUE,
  settlement_age = 0
)
}
\arguments{
\item{file_path}{A file path to a directory where the assessment input and output files will be saved.}

\item{r0}{Unexploited recruitment from Beverton-Holt stock-recruitment model.}

\item{r0_phase}{Phase for Unexploited recruitment from Beverton-Holt stock-recruitment model. Default value is 1.}

\item{steepness}{Steepness from Beverton-Holt stock-recruitment model. Recruitment relative to unfished recruitment at 20\% of unfished spawning biomass.}

\item{sigmar}{Standard deviation of log recruitment.}

\item{projection_f}{Fishing mortality inputs for projections.}

\item{projection_catch}{Catch inputs for projections.}

\item{sa_data}{Simulated stock assessment input data from Rscript/simulation.R.}

\item{model_year}{A vector of years for model fitting.}

\item{projection_year}{A vector of years for projection.}

\item{use_depletion}{A logical value.}

\item{depletion_ratio}{Ratio of depletion.}

\item{initial_equilibrium_catch}{TRUE or FALSE.}

\item{settlement_age}{Settlement age. Default value is age 0.}
}
\value{
Outputs from Stock Synthesis 3
}
\description{
Generate frequently used input data for Stock Synthesis 3
}
\examples{
\dontrun{
generate_ss3(
  file_path  = here::here("data", "data_rich"),
  r0 = 12, steepness = 0.5, sigmar =0.25,
  projection_f = 0.75, projection_catch = NULL,
  sa_data = sa_data, model_year = 1985:2012,
  projection_year = 2013:2017,
  use_depletion = FALSE, depletion_ratio = NULL
)
}
}
