% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic}
\alias{logistic}
\title{Compute various logistic patterns}
\usage{
logistic(
  pattern,
  x,
  slope_asc = NULL,
  location_asc = NULL,
  slope_desc = NULL,
  location_desc = NULL
)
}
\arguments{
\item{pattern}{logistic function option. Options: "simple_logistic" and "double_logistic"}

\item{x}{Vector or scalar of indepenent variable}

\item{slope_asc}{Slope parameter for ascending curve}

\item{location_asc}{Location parameter for ascending curve}

\item{slope_desc}{Slope parameter for descending curve}

\item{location_desc}{Location parameter for descending curve}
}
\value{
Values of function at \code{x}.
}
\description{
Compute various logistic patterns
}
\examples{
\dontrun{
logistic(pattern = "simple_logistic", x = c(1:12), slope_asc = 3, location_asc = 1.5)
logistic(pattern = "double_logistic", x = c(1:12), slope_asc = 3, location_asc = 1.5, slope_desc = 2, location_desc = 10)
}
}
