% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ewe_output.R
\name{read_ewe_effort}
\alias{read_ewe_effort}
\title{Read in EwE fishing effort data and make figures}
\usage{
read_ewe_effort(
  file_path,
  file_names,
  skip_nrows = 8,
  colname_1 = "month",
  fleets,
  years
)
}
\arguments{
\item{file_path}{a character string shows path to the working folder where EwE output csv files are located.}

\item{file_names}{a vector of csv file names.}

\item{skip_nrows}{integer: the number of lines of the data file to skip before reading data.}

\item{colname_1}{column name of the first column. Default name is Month.}

\item{fleets}{a character string describes the column of the data matrix, which is fleets in EwE case.}

\item{years}{a vector of years.}
}
\description{
It reads in the output data from Ecopath with Ecosim (EwE) and
making figures for each dataset. The timestep of the output data needs to be
month. The rows of the data matrix represents month and the columns of the
data represents fleets. It creates figures and save the figures where EwE
output csv files are located. The potential data include: fishing effort
over years.
}
