% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ewe_output.R
\name{read_ewe_output}
\alias{read_ewe_output}
\title{Read in EwE output data and make figures}
\usage{
read_ewe_output(
  file_path,
  file_names,
  skip_nrows = 8,
  functional_groups,
  plot = FALSE,
  figure_titles = NULL,
  figure_colors = rainbow(12)
)
}
\arguments{
\item{file_path}{a character string shows path to the working folder where EwE output csv files are located.}

\item{file_names}{a vector of csv file names.}

\item{skip_nrows}{interger: the number of lines of the data file to skip before reading data.}

\item{functional_groups}{a character string describes the column of the data matrix, which is functional groups in EwE case.}

\item{plot}{logical: if \code{TRUE} then make figures for all output data files.}

\item{figure_titles}{a character vector of strings describes the names of each figure.}

\item{figure_colors}{a vector of colors.}
}
\description{
It reads in the output data from Ecopath with Ecosim (EwE) and making figures for each dataset. The timestep of the output data needs to be year. The rows of the data matrix represents years and the columns of the data represents functional groups. It creates figures and save the figures where EwE output csv files are located. The potential data include: biomass, catch, consumption biomass, feeding time, mortality, trophic level, and average weight (or production to comsumption) over years.
}
