% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ewe_output.R
\name{read_ewe_reference_points}
\alias{read_ewe_reference_points}
\title{Read in EwE MSY reference points and make figures}
\usage{
read_ewe_reference_points(
  file_path,
  file_names,
  functional_groups,
  key_functional_group,
  ages,
  reference_points_scenario,
  plot = FALSE
)
}
\arguments{
\item{file_path}{a character string shows path to the working folder where EwE output csv files are located.}

\item{file_names}{a vector of csv file names.}

\item{functional_groups}{a character string describes the column of the data matrix, which is functional groups in EwE case.}

\item{key_functional_group}{a character string describes the key functional group.}

\item{ages}{age classes of the functional groups.}

\item{reference_points_scenario}{a character string describes the reference points scenario: compensation or stationary.}

\item{plot}{logical: if \code{TRUE} then make figures for all output data files.}
}
\description{
It reads in the MSY reference points (i.e. MSY, FMSY, BMSY) from Ecopath with Ecosim (EwE) and making figures for yield over fishing mortality. It only works with Atlantic menhaden for now.
}
