% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ewe_output.R
\name{read_ewe_reference_points_vec}
\alias{read_ewe_reference_points_vec}
\title{Extract EwE MSY reference points using catch over fishing effort multiplier table}
\usage{
read_ewe_reference_points_vec(
  file_path,
  file_names,
  functional_groups,
  key_functional_group,
  ages,
  reference_points_scenario
)
}
\arguments{
\item{file_path}{a character string shows path to the working folder where EwE reference points output csv files are located.}

\item{file_names}{a csv file names.}

\item{functional_groups}{a character string describes the column of the data matrix, which is functional groups in EwE case.}

\item{key_functional_group}{a character string describes the key functional group.}

\item{ages}{age classes of the functional groups.}

\item{reference_points_scenario}{a character string describes the reference points scenario: compensation or stationary.}
}
\description{
Extract MSY reference points (i.e. MSY, FMSY, BMSY) from Ecopath with Ecosim (EwE).
}
