% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ewe_output.R
\name{read_fmsy_table}
\alias{read_fmsy_table}
\title{Read in EwE MSY reference points summary table}
\usage{
read_fmsy_table(
  file_names,
  skip_nrows = 12,
  colnames = c("Group", "TL", "Fbase", "Cbase", "Vbase", "FmsyFound", "Fmsy", "Cmsy",
    "Vmsy", "CmsyAll", "VmsyAll"),
  functional_groups
)
}
\arguments{
\item{file_names}{a csv file names.}

\item{skip_nrows}{integer: the number of lines of the data file to skip before reading data.}

\item{colnames}{column names of the csv file.}

\item{functional_groups}{a character string describes the column of the data matrix, which is functional groups in EwE case.}

\item{key_functional_group}{a character string describes the key functional group.}
}
\description{
It reads in the MSY reference points (i.e. MSY, FMSY, BMSY) from Ecopath with Ecosim (EwE).
}
