% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishery.R
\name{create_fishery}
\alias{create_fishery}
\title{Create fishery from Ecopath with Ecosim (EwE) output data}
\usage{
create_fishery(
  file_path,
  skip_nrows,
  species,
  species_labels,
  years,
  fleet_num,
  selectivity = NULL,
  CV,
  sample_num,
  waa_path
)
}
\arguments{
\item{file_path}{A character string shows path to Ecosim catch output file (e.g., catch_annual.csv) is located.}

\item{skip_nrows}{interger; the number of lines of the data file to skip before reading data.}

\item{species}{The species to create fishery for. A vector of column names from the catch_annual.csv.}

\item{species_labels}{Customized names for speicies columns.}

\item{years}{A vector of years of the fishery.}

\item{fleet_num}{Fleet number.}

\item{selectivity}{Selectivity at age. A dataframe defining selectivity-at-age for a species over time. Columns represent age classes and rows represent years. Provide selectivity data only if \code{fleet_num} is greater than 1.}

\item{CV}{CV of catch. A vector of CVs for individual fleets.}

\item{sample_num}{Annual sample size. A vector of sample sizes for indiviudal fleets.}

\item{waa_path}{A character string shows path to Ecosim weight output file (e.g., weight_annual.csv) is located.}
}
\value{
A data list that includes catch and composition data from the EwE operating model, and observed catch and composition data for each fleet.
Each fleet has time series catch with lognormal error and composition data with sampling error.
}
\description{
This function creates fishery observations for an Ecosim scenario.
}
\details{
The function takes fishery catch data from an Ecosim scenario.
The catch data are read in from Ecosim catch_annual.csv and the catch data are in biomass.
The function reads in total catch-at-age data of a specific species by age classes and time.
EwE already applies a fishery selectivity internally, so no need to reallocate catch if there is only one fleet for the fishery.
}
\concept{create_data functions}
